+++ lib/dotcom/utils/service_date_time.ex
@@ -1,29 +1,18 @@
 defmodule Dotcom.Utils.ServiceDateTime do
   @moduledoc """
-  A collection of functions that helps to work with date_times with regard to service ranges.
-  Currently, we consider the most general case where service starts at 03:00:00am and ends at 02:59:59am.
-
-  In the future, we aim to add route-specific service times.
-
-  The service range continuum:
-
-  <---before today---|---later this week---|---next week---|---after next week--->
-                   today
-
-  Before today and after next week are open intervals.
+  TODO: Add module documentation
   """
 
   require Logger
 
   use Timex
 
+  import Dotcom.Utils.DateTime, only: [coerce_ambiguous_time: 1, now: 0, timezone: 0]
+
   alias Dotcom.Utils
 
-  @type named_service_range() ::
-          :before_today | :today | :later_this_week | :next_week | :after_next_week
-  @date_time_module Application.compile_env!(:dotcom, :date_time_module)
   @service_rollover_time Application.compile_env!(:dotcom, :service_rollover_time)
-  @timezone Application.compile_env!(:dotcom, :timezone)
+  @timezone timezone()
 
   @doc """
   Returns the time at which service rolls over from 'today' to 'tomorrow'.
@@ -32,41 +21,16 @@ defmodule Dotcom.Utils.ServiceDateTime do
 
   @doc """
   Get the service date for the given date_time.
-  If the time is before 03:00:00am, we consider it to be the previous day.
+  If the time is before 3am, we consider it to be the previous day.
   """
   @spec service_date() :: Date.t()
   @spec service_date(DateTime.t()) :: Date.t()
-  def service_date(date_time \\ @date_time_module.now()) do
+  def service_date(date_time \\ now()) do
     if date_time.hour < @service_rollover_time.hour do
-      Timex.shift(date_time, hours: -@service_rollover_time.hour)
-      |> @date_time_module.coerce_ambiguous_date_time()
-      |> Timex.to_date()
+      Timex.shift(date_time, hours: -@service_rollover_time.hour) |> Timex.to_date()
     else
       Timex.to_date(date_time)
     end
-  end
-
-  @doc """
-  The service range for the given date_time.
-  """
-  @spec service_range(DateTime.t()) :: named_service_range()
-  def service_range(date_time) do
-    Enum.find(
-      [
-        &service_before_today?/1,
-        &service_today?/1,
-        &service_later_this_week?/1,
-        &service_next_week?/1,
-        &service_after_next_week?/1
-      ],
-      fn f ->
-        f.(date_time)
-      end
-    )
-    |> Kernel.inspect()
-    |> Kernel.then(fn module -> Regex.run(~r/_(\w+)\?/, module) end)
-    |> List.last()
-    |> String.to_atom()
   end
 
   @doc """
@@ -74,11 +38,11 @@ defmodule Dotcom.Utils.ServiceDateTime do
   """
   @spec beginning_of_next_service_day() :: DateTime.t()
   @spec beginning_of_next_service_day(DateTime.t()) :: DateTime.t()
-  def beginning_of_next_service_day(datetime \\ @date_time_module.now()) do
+  def beginning_of_next_service_day(datetime \\ now()) do
     datetime
     |> end_of_service_day()
     |> Timex.shift(microseconds: 1)
-    |> @date_time_module.coerce_ambiguous_date_time()
+    |> coerce_ambiguous_time()
   end
 
   @doc """
@@ -86,11 +50,11 @@ defmodule Dotcom.Utils.ServiceDateTime do
   """
   @spec beginning_of_service_day() :: DateTime.t()
   @spec beginning_of_service_day(DateTime.t()) :: DateTime.t()
-  def beginning_of_service_day(date_time \\ @date_time_module.now()) do
+  def beginning_of_service_day(date_time \\ now()) do
     date_time
     |> service_date()
     |> Timex.to_datetime(@timezone)
-    |> @date_time_module.coerce_ambiguous_date_time()
+    |> coerce_ambiguous_time()
     |> Map.put(:hour, @service_rollover_time.hour)
   end
 
@@ -99,23 +63,22 @@ defmodule Dotcom.Utils.ServiceDateTime do
   """
   @spec end_of_service_day() :: DateTime.t()
   @spec end_of_service_day(DateTime.t()) :: DateTime.t()
-  def end_of_service_day(date_time \\ @date_time_module.now()) do
+  def end_of_service_day(date_time \\ now()) do
     date_time
     |> service_date()
     |> Timex.to_datetime(@timezone)
-    |> @date_time_module.coerce_ambiguous_date_time()
+    |> coerce_ambiguous_time()
     |> Timex.shift(days: 1, hours: @service_rollover_time.hour, microseconds: -1)
-    |> @date_time_module.coerce_ambiguous_date_time()
+    |> coerce_ambiguous_time()
     |> Map.put(:hour, @service_rollover_time.hour - 1)
   end
 
   @doc """
   Get a service range for the day of the given date_time.
-  Service days go from 03:00:00am to 02:59:59am the following day.
   """
-  @spec service_range_day() :: Utils.DateTime.date_time_range()
-  @spec service_range_day(DateTime.t()) :: Utils.DateTime.date_time_range()
-  def service_range_day(date_time \\ @date_time_module.now()) do
+  @spec service_range_day() :: Utils.DateTime.time_range()
+  @spec service_range_day(DateTime.t()) :: Utils.DateTime.time_range()
+  def service_range_day(date_time \\ now()) do
     beginning_of_service_day = beginning_of_service_day(date_time)
     end_of_service_day = end_of_service_day(date_time)
 
@@ -124,55 +87,62 @@ defmodule Dotcom.Utils.ServiceDateTime do
 
   @doc """
   Get a service range for the week of the given date_time.
-  Service weeks go from Monday at 03:00:00am to the following Monday at 02:59:59am.
-  If today is the last day in the service week, the range will be the same as the range for today.
   """
-  @spec service_range_later_this_week() :: Utils.DateTime.date_time_range()
-  @spec service_range_later_this_week(DateTime.t()) :: Utils.DateTime.date_time_range()
-  def service_range_later_this_week(date_time \\ @date_time_module.now()) do
-    beginning_of_next_service_day = beginning_of_next_service_day(date_time)
+  @spec service_range_current_week() :: Utils.DateTime.time_range()
+  @spec service_range_current_week(DateTime.t()) :: Utils.DateTime.time_range()
+  def service_range_current_week(date_time \\ now()) do
+    beginning_of_current_week =
+      date_time
+      |> Timex.beginning_of_week()
+      |> Timex.shift(hours: 24)
+      |> beginning_of_service_day()
 
-    end_of_later_this_week = date_time |> Timex.end_of_week() |> end_of_service_day()
+    end_of_current_week = date_time |> Timex.end_of_week() |> end_of_service_day()
 
-    case Timex.compare(beginning_of_next_service_day, end_of_later_this_week) do
-      1 -> service_range_day(date_time)
-      _ -> {beginning_of_next_service_day, end_of_later_this_week}
-    end
+    {beginning_of_current_week, end_of_current_week}
   end
 
   @doc """
   Get a service range for the week following the current week of the given date_time.
   """
-  @spec service_range_next_week() :: Utils.DateTime.date_time_range()
-  @spec service_range_next_week(DateTime.t()) :: Utils.DateTime.date_time_range()
-  def service_range_next_week(date_time \\ @date_time_module.now()) do
-    {_, end_of_later_this_week} = service_range_later_this_week(date_time)
-    beginning_of_next_week = Timex.shift(end_of_later_this_week, microseconds: 1)
+  @spec service_range_following_week() :: Utils.DateTime.time_range()
+  @spec service_range_following_week(DateTime.t()) :: Utils.DateTime.time_range()
+  def service_range_following_week(date_time \\ now()) do
+    {_, end_of_current_week} = service_range_current_week(date_time)
+    beginning_of_following_week = Timex.shift(end_of_current_week, microseconds: 1)
 
-    end_of_next_week =
-      beginning_of_next_week |> Timex.end_of_week() |> end_of_service_day()
+    end_of_following_week =
+      beginning_of_following_week |> Timex.end_of_week() |> end_of_service_day()
 
-    {beginning_of_next_week, end_of_next_week}
+    {beginning_of_following_week, end_of_following_week}
   end
 
   @doc """
-  Get a service range for all time after the following week of the given date_time.
+  Get a service range for all time afer the following week of the given date_time.
   """
-  @spec service_range_after_next_week() :: Utils.DateTime.date_time_range()
-  @spec service_range_after_next_week(DateTime.t()) :: Utils.DateTime.date_time_range()
-  def service_range_after_next_week(date_time \\ @date_time_module.now()) do
-    {_, end_of_next_week} = date_time |> service_range_next_week()
-    beginning_of_after_next_week = Timex.shift(end_of_next_week, microseconds: 1)
+  @spec service_range_later() :: Utils.DateTime.time_range()
+  @spec service_range_later(DateTime.t()) :: Utils.DateTime.time_range()
+  def service_range_later(date_time \\ now()) do
+    {_, end_of_following_week} = date_time |> service_range_following_week()
+    beginning_of_later = Timex.shift(end_of_following_week, microseconds: 1)
 
-    {beginning_of_after_next_week, nil}
+    {beginning_of_later, nil}
   end
 
   @doc """
-  Is the given date_time before the beginning of service today?
+  Given a time_range and a date_time, returns true if the date_time is within the time_range.
   """
-  @spec service_before_today?(DateTime.t()) :: boolean
-  def service_before_today?(date_time) do
-    Timex.before?(date_time, beginning_of_service_day())
+  @spec in_range?(Utils.DateTime.time_range(), DateTime.t()) :: boolean
+  def in_range?({nil, stop}, date_time) do
+    Timex.before?(date_time, stop) || Timex.equal?(date_time, stop)
+  end
+
+  def in_range?({start, nil}, date_time) do
+    Timex.after?(date_time, start) || Timex.equal?(date_time, start)
+  end
+
+  def in_range?({start, stop}, date_time) do
+    in_range?({start, nil}, date_time) && in_range?({nil, stop}, date_time)
   end
 
   @doc """
@@ -180,15 +150,15 @@ defmodule Dotcom.Utils.ServiceDateTime do
   """
   @spec service_today?(DateTime.t()) :: boolean
   def service_today?(date_time) do
-    service_range_day() |> @date_time_module.in_range?(date_time)
+    service_range_day() |> in_range?(date_time)
   end
 
   @doc """
   Does the given date_time fall within the service range of this week?
   """
-  @spec service_later_this_week?(DateTime.t()) :: boolean
-  def service_later_this_week?(date_time) do
-    service_range_later_this_week() |> @date_time_module.in_range?(date_time)
+  @spec service_this_week?(DateTime.t()) :: boolean
+  def service_this_week?(date_time) do
+    service_range_current_week() |> in_range?(date_time)
   end
 
   @doc """
@@ -196,14 +166,14 @@ defmodule Dotcom.Utils.ServiceDateTime do
   """
   @spec service_next_week?(DateTime.t()) :: boolean
   def service_next_week?(date_time) do
-    service_range_next_week() |> @date_time_module.in_range?(date_time)
+    service_range_following_week() |> in_range?(date_time)
   end
 
   @doc """
   Does the given date_time fall within the service range after next week?
   """
-  @spec service_after_next_week?(DateTime.t()) :: boolean
-  def service_after_next_week?(date_time) do
-    service_range_after_next_week() |> @date_time_module.in_range?(date_time)
+  @spec service_later?(DateTime.t()) :: boolean
+  def service_later?(date_time) do
+    service_range_later() |> in_range?(date_time)
   end
 end
